Series of refactors
* Split solids into files by pipeline, with potential for common solid files for shared solids
* Add secret config to helmfile
* Use configured decorator to pre-configure static settings for solids/resources
