+++ orchestration/dagster_orchestration/.env.test
@@ -1,0 +1,5 @@
+HCA_ARGO_URL=https://argo.zombo.com.fake/
+DATA_REPO_URL=https://datarepo.zombo.com.fake/
+HCA_GOOGLE_PROJECT=fake-google-project
+HCA_GCP_ENV=dev
+SLACK_NOTIFICATIONS_CHANNEL="#beepbloopfakechannel"
