+++ orchestration/dagster_orchestration/hca_manage/main.py
@@ -5,7 +5,7 @@ from data_repo_client import ApiClient, Configuration, RepositoryApi
 
 from hca_manage import __version__ as hca_manage_version
 from .manage import HcaManage
-from hca_orchestration.resources.base import default_google_access_token
+from hca_orchestration.contrib.google import default_google_access_token
 
 data_repo_host = {
     "dev": "https://jade.datarepo-dev.broadinstitute.org/",
