+++ orchestration/dagster_orchestration/hca_orchestration/environments/test_validate_egress.yaml
@@ -1,6 +1,8 @@
 solids:
   post_import_validate:
     config:
-      gcp_env: dev
-      google_project_name: fake-google-project
       dataset_name: hca_aeiou
+  notify_slack_of_egress_validation_results:
+    config:
+      dataset_name: hca_aeiou
+
