+++ orchestration/dagster_orchestration/hca_orchestration/pipelines/validate_egress.py
@@ -1,16 +1,14 @@
 from dagster import ModeDefinition, pipeline
 
-from dagster_slack import slack_resource
-
-from hca_orchestration.resources import jade_data_repo_client, noop_data_repo_client, console_slack_client
-from hca_orchestration.solids import post_import_validate, notify_slack_of_egress_validation_results
+from hca_orchestration.resources import jade_data_repo_client, noop_data_repo_client, console_slack_client, live_slack_client
+from hca_orchestration.solids.validate_egress import post_import_validate, notify_slack_of_egress_validation_results
 
 
 prod_mode = ModeDefinition(
     name="prod",
     resource_defs={
         "data_repo_client": jade_data_repo_client,
-        "slack": slack_resource,
+        "slack": live_slack_client,
     }
 )
