+++ orchestration/dagster_orchestration/hca_orchestration/resources/__init__.py
@@ -1,6 +1,6 @@
 from hca_orchestration.resources.beam import dataflow_beam_runner, local_beam_runner, test_beam_runner
 from hca_orchestration.resources.data_repo import jade_data_repo_client, noop_data_repo_client
-from hca_orchestration.resources.slack import console_slack_client
+from hca_orchestration.resources.slack import console_slack_client, live_slack_client
 from hca_orchestration.resources.storage import google_storage_client, local_storage_client
 
 
@@ -9,6 +9,7 @@ __all__ = [
     dataflow_beam_runner,
     google_storage_client,
     jade_data_repo_client,
+    live_slack_client,
     local_beam_runner,
     local_storage_client,
     noop_data_repo_client,
