+++ orchestration/dagster_orchestration/hca_orchestration/resources/data_repo.py
@@ -1,4 +1,6 @@
-from dagster import resource, StringSource, Field
+import os
+
+from dagster import configured, resource, StringSource, Field
 
 from data_repo_client import ApiClient, Configuration, RepositoryApi
 
@@ -8,7 +10,7 @@ from hca_orchestration.contrib.google import default_google_access_token
 @resource({
     "api_url": Field(StringSource)
 })
-def jade_data_repo_client(init_context):
+def base_jade_data_repo_client(init_context):
     # create API client
     config = Configuration(host=init_context.resource_config["api_url"])
     config.access_token = default_google_access_token()
@@ -17,6 +19,13 @@ def jade_data_repo_client(init_context):
 
     # submit file ingest (for now just enumerate datasets or something to prove interaction works)
     return RepositoryApi(api_client=client)
+
+
+@configured(base_jade_data_repo_client)
+def jade_data_repo_client(config):
+    return {
+        'api_url': os.environ.get('DATA_REPO_URL'),
+    }
 
 
 @resource
