+++ orchestration/dagster_orchestration/mypy.ini
@@ -21,6 +21,9 @@ ignore_missing_imports = True
 [mypy-dagster_k8s.*]
 ignore_missing_imports = True
 
+[mypy-dagster_slack.*]
+ignore_missing_imports = True
+
 [mypy-data_repo_client.*]
 ignore_missing_imports = True
