+++ orchestration/dagster_orchestration/pytest.ini
@@ -1,2 +1,4 @@
 [pytest]
-ENV=test
+console_output_style = progress
+env_files =
+	.env.test
